Move call to cleanRides function to before managing rides
It's possible that the user tries to update a ride that has already passed. In those cases, we do not want to update it on the db, because it would lead to a scenario of an empty document that was causing problems whenever listing the rides. Therefore, we chose to clean the rides first, changing the message to the user accordingly, in case of a ride that no longer exists