Hotfix: undefined user when listing rides (#36)
* Move call to cleanRides function to before managing rides

It's possible that the user tries to update a ride that has already passed. In those cases, we do not want to update it on the db, because it would lead to a scenario of an empty document that was causing problems whenever listing the rides. Therefore, we chose to clean the rides first, changing the message to the user accordingly, in case of a ride that no longer exists

* Make small changes to async calls

The callbacks from the connect and disconnect functions from the MongoDB package are about to be deprecated. Therefore, we're using methods to log whether it failed or not.

* Return rides to be removed on cleanRides

Used to check whether the ride that the person is trying to modify was already removed or not.

* Fix message for setting rides as full

* Add await to avoid executing function out of order